修改验证码发送的点击限制 修改工程信息地图页的权限请求，增加gps开关检测 修改工程信息支付界面的信息共享问题 修改折线图y轴数值限制，>0 修改工程发布联系人公司信息传递问题 新增工程圈底部按钮和fragment

+++ app/src/main/java/com/shuangduan/zcy/model/bean/ProjectSubViewBean.java
@@ -121,7 +121,7 @@ public class ProjectSubViewBean {
          */
 
         private String time;
-        private int value;
+        private float value;
 
         public String getTime() {
             return time;
@@ -131,11 +131,11 @@ public class ProjectSubViewBean {
             this.time = time;
         }
 
-        public int getValue() {
+        public float getValue() {
             return value;
         }
 
-        public void setValue(int value) {
+        public void setValue(float value) {
             this.value = value;
         }
     }
